Fix Floating Point Comparison Issue
GDAL GetNoDataValue() returns a double, which can be problematic when
comparing against pixel values stored in lower precision formats, e.g.
float32, float16. This bug manifested when some rasters were found to
have -3.4028235e+38 set as nodatavalue. -3.4028235e+38 is the smallest
float32 value, which does not equal the 'same' value in double precision.
We now cast nodatavalue to the same format as the pixel values.