Fix Spatial Shift
We now decompose and reconstruct images based on wavedec2 and waverec2,
respectively. This replaces the earlier approach which relied on dwt2
and idwt2, where a spatial shift was being introduced during the
reconstruction stage -- although, not sure why that was the case. It
could be an undocumented feature/bug in pywt.