Fix git hash in metadata
Git hash gets taken in setup.py and written as a Python module. This is so the hash reflects the currently installed version of UncoverML. Problem is that the git_hash module seems to be written after modules are gathered as part of install, so it can't be imported. Instead of relying on writing to a file, I've changed the version number to be 'major.minor.patch-githash' which gets parsed during setup. This version number now gets included in metadata.


Former-commit-id: 0ba7b8d104f4289c09e3463c83d8fbe4eb96598a