Added 'index' and 'prediction' columns to rawcovrates CSV
These columns have been added on request. The 'index' column is
intended to be a UID for post-processing, and the 'prediction' column
is the cross-val predicted value for that location.

These get added as part of the 'save_intersected...' function. If an
'index' field exists in target shapefile, it will be used to populate
the column. If not, it will be zero filled. If cross-val is run, the
'export_crossval' (refactored from geoio.py to be part of validate.py)
will populate the 'prediction' column. If it is not run, it will be
zero filled.


Former-commit-id: f6855dec692c6400c6ea7ec107f4795105751be1