Start of target search implementation
The goal here is to train a classifier on two groups of points -
class A being randomly sampled in a prediction area that is a subset
of a dataset, and class B being targets from the full set of data.

Class B points that are classified as being likely to belong to class A
are then included in the training targets for a subsequent model. The
idea is to make a model more robust by including like data from outside
a prediction area.

- Add target search script and target search crop box (prediction area)
- Generate random samples in prediction area
- Merge with targets from wider training set
- Train and predict a classification

Next step is to modify prediction so we can classify on just the target
points instead of the entire area of the provided data.