WIP: return original target data
Target search currently returns selected targets with their
observations replaced by classification labels. This makes them not
useful for training other models.

Need to turn the targets with their original observation data, which
is trickier than it seems. The idea here (which doesn't work) is to
store a copy of the targets with the original data before they get
labelled, then at the end use the 'selected targets' to index filter
the original data.

However, targets need to be gathered across all processors, which
requires a 'keep' mask which is generated by the feature transform
function, which is dependent on the targets being used, so we have to
do a full set of feature transforms if we want to handle multiple
Target objects, and this and other issues makes it unnessecarily
complicated.

A simpler approach which I'll try next is to copy the original data
to a 'fields' entry on the Targets object, carry the one Targets object
through the script and then retrieve the original data from the field
at the end.


Former-commit-id: 64c2f4f893739a884e27e3d8ae85330fec3dfc8f