Sort targets by position when merging
This fixes the discrepancy in results between runs with different
numbers of processors.

Uncoverml feature extraction requires that targets be sorted
lexically by position (with positions in 'y, x' ordering).

Assumption was that it intersected targets and covariates and delivered
them in whatever order the targets were in when inputted. This isn't
the case, and targets not in the 'canonical' ordering will cause
mismatches between covariate data and target positions when using
more than one processor (due to image chunking).