Add training of bootstrap models
Can now train an ensemble of models on bootstrapped data, which will
allow for pseudo-probabilistic predictions.

Straightforward implementation. If bootstrap is set in config, learning
will resample training data N times, training N models.
Targets/features are only intersected once, with different combinations
of data being reused for training depending on the result of
resampling.

If bootstrapping is enabled, learn will output a list of models. These
will still be stored in the .model as usual, but now a check has to be
done when unpickling whether the model is ensemble or not.