Modify config module and CLI commands to fix tests
Changes to the config module (trying to make it more flexible,
require less unneeded params based on command being run) broke tests.
Have had to make changes rectifying this.

The major one is detecting algorithm based on model being used for
prediction, because algo property of config is used, but I don't want
users to have to provide the algorithm in the config (both out of
convenience, but also if you get the .model from someone else you may
not know the algorithm).

The config is getting unwieldy, so it may be time to refactor to have
a different parsing method for each command. Would also make config
parsing much easier to test.


Former-commit-id: 4b3d7e57e1ecf878c0b25503899a332aff738b12