Fix bug in gathering targets and features
`gather_targets` and `gather_features` are meant to collect
target/feature fragments from across nodes onto a single node.
It gets called with `node=0`, e.g. gather the data onto node 0.
However, there's a condition `if node` then gather onto that node,
otherwise do an `allgather` and duplicate across all nodes
If we call with `node=0` then it doesn't gather onto node 0,
but duplicates on all nodes, because 0 evaluates to False.

The code in the first branch (if node...) also failed as it was trying
to concatenate directly on the result of `comm.gather`, which will
return None if not on the node being gathered to. So nodes that aren't
root will crash.

Fixing this bug will break multirandomforest, it currently relies
on training data being duplicated on all nodes. We'll have to change
to explicitly duplicate data when running this algorithm, or extend
the shared memory implmentation currently being built for bootstrapping
to work with multirandomforest.


Former-commit-id: cb3a9ae99ebc5a972b078b22200e867334bf9f62