Make multirandomforest compatible with shared mem
MRF relied on training data being duplciated on all processes (which
was the result of a bug). Now that the bug is fixed, MRF needs
shared memory (or deliberate duplication, but we another option now)
to work.

To allow MRF to work, and any other parallel model in future (e.g.
bootstrapping), training data is now placed in shared arrays at
top-level scope following target/covariate intersection.

This does mean the shared memory in the bootstrap learn function is now
redundant and needs to be refactored.