Optimise bootstrap crossval
If crossvalidation is run sequentially, allow bootstrap SVR to use
all available processors to train submodels. This is useful if there
are more CPUs then folds being trained as it allows validation to
utilise 100% of available CPUs in all situations.

This should be extended to apply to parallel models
(e.g. multirandomforest).


Former-commit-id: 7b8a6d644f62cc606c72674a28a46a62f981fd86