Allow user-define scorers for GridSearchCV
This allows users to specify a list of metrics to be used with
hyperparameter tuning.

The implmentation involves wrapping sklearn/revrand metrics with
sklearn's `make_scorer` method. Currently metrics that require a
`predict_dist` method and metrics that output collections are not
supported.

If scorers are not provided, then the default `score` method of the
model is used.

Also update scikit-learn to 0.22.2.