Fix: handle null rows when adding Y to RCV table
When exporting crossvalidation results, we add the predicted values from
cross-validation to the rawcovariates table. If any of the target points
are dropped due to having all NaN covariate values, then these will not
be included when predicting during cross-validation. Thus there will be
a mismatch between length of predictions and rows in rawcovariates table
which raises an exception.

To get around this, load the rawcovariates mask and fill the prediction
array with zeros where dropped rows exist.