Fix covariate transforms for targetsearch
Previous behaviour was to save intersected covariate/target data
after targetsearch and load it in the learn step after the provided
targets/covariates had been intersected and covariates transformed,
and append the data to the end of that.

This is incorrect as we should be intersecting the data and performing
transforms on the full set of data (provided shapefile + targets from
targetsearch).

So new behaviour is to only save targets from targetsearch (not
covariate data as well - not sure why I did that. Probably thought I was
clever by not having to repeat the transform), load them in learn
and merge them with shapefile targets, and then intersect with
covariates and apply transforms.


Former-commit-id: 85fea75b87e298f3823162a78f4dc97e657cef15