Only replace transforms rather than whole FS
Feature set has paths baked in, so when running tests on a different
platform (circlCI), the predict step breaks because it loads the
presaved feature sets with paths from my machine.

By just reassigning the saved transform, we avoid this. A better fix is
to add the 'name' parameter back to feature sets and save transforms as
a dict of [name: transform] and assign pickled transforms based on this.
This also allows people to use new covariates in the predict step and
assign them a transform.