Fix cropping/target filtering bug
As part of cropping, targets not within the image are filtered. This
also applies if cropping is done outside uncoverml and the covariate
image bounds are greater than the target bounds.

Coordinates are filtered out if they are not within image bounds.
They are then ordered by Y, X and this ordering is used to select the
correpsonding observation and field target values. The bug is that
because the coord array has already been filtered, the value arrays and
coord arrays have different lengths, so the ordered index doesn't
correspond to the correct observation values.

To fix this we now filter the coordinates and value arrays at the same
time when loading the shapefile.


Former-commit-id: 921fdcbb078fe8e90fce83f9f07a21c7f9118879