Add sample weighting
Weights can now be provided for models that support 'sample_weight'
argument for their fit method.

The weights need to exist as a field on the shapefile points, and the
name of this field is provided as 'weight_property' parameter in the
config under 'targets' block. If this parameter is provided, and the
model supports sample weighting, then the sample weights will be applied
when fitting the model.