Add arbitrary fields to rawcovariates csv
A list of shapefile field names specified in config as 'write_to_csv'
parameter (e.g. wrte_to_csv: [SITEID, type]) will be written to the
rawcovariates CSV file. This can be used for debugging, validation and
so on. If any of the provided field names don't exist in the shapefile,
an error will be thrown.


Former-commit-id: 82c2e1f1c88dff6bc69220a006617e7d7c22ac38