Support for training on features from target file
Can now set source of features as 'table', specifying a list of
'fields'. These are columns in the target shapefile that will be
read as covaraites. Can also specify a 'ndv' parameter under 'table'
for speciying the NDV of the shapefile (these will be masked out).

A 'drop' field is added to 'targets'. Rows where the target property
is equal to this value will be dropped when loading targets, such as
when there are unknown values in the property column intended to be
predicted later. This needs to be fixed as currently the rows will be
dropped then covariate data will be taken, so the unknown rows won't
be predicted on.