Fix strange 0 variance bug
In the course of developing support for tabular data, some predictions
using random forest gave a variance of 0 when using predict dist.
This then breaks the _normpdf of the quadrature function (divides by
std, if variance is 0 then std is 0).

This is occurring somewhere in revrand - their predict_moments
function is very rarely outputting an expected variance of 0. This can
be seen by running learn on the 'tabular_prediction' config with a
random state of 1 for the randomforest.

A workaround is to stabilise with a very small value if the std is 0.


Former-commit-id: f72feab91d4df413629357062590157d826185db