Read tabular features from own shapefile
Rather than piggybacking on the target shapefile, load features from
a shapefile specified in features block. This allows us to use features
and targets from different shapefiles.

It also makes the issue of dropping rows from the target shapefile
easier, as we can drop the targets and create a mask for intesecting
with the features so we don't train on those rows, but still provide
the full shapefile for prediction.


Former-commit-id: 121efba0a9e36fcc927be161995742de37a7d670